refactor(desktop): extract assistant turn display logic
